pyproject.toml
[project]
name = "repl-nix-difyapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "scikit-learn>=1.5.2",
    "python-dotenv>=1.0.1",
    "mecab-python3>=1.0.10",
    "unidic-lite>=1.0.8",
    "python-mecab-ko",
]
